pyproject.toml
[project]
name = "httpt"
version = "0.3.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cachetools>=6.2.4",
    "circuitbreaker>=2.1.3",
    "fastapi>=0.127.0",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "requests>=2.32.5",
    "tiktoken>=0.12.0",
    "transformers>=4.57.3",
    "uvicorn>=0.40.0",
]
